move console handler.
