+++ packages/cli/src/index.js
@@ -5,8 +5,30 @@ import path from 'path'
 import { config } from 'dotenv-defaults'
 import yargs from 'yargs'
 
-import { getPaths, getConfigPath } from '@redwoodjs/internal'
-import { telemetryMiddleware } from '@redwoodjs/telemetry'
+// import { getPaths, getConfigPath } from '@redwoodjs/internal'
+// import { telemetryMiddleware } from '@redwoodjs/telemetry'
+
+import * as buildCommand from './commands/build'
+// import * as checkCommand from './commands/check'
+import * as consoleCommand from './commands/console'
+// import * as dataMigrateCommand from './commands/data-migrate'
+// import * as deployCommand from './commands/deploy'
+// import * as destroyCommand from './commands/destroy'
+import * as devCommand from './commands/dev'
+// import * as execCommand from './commands/exec'
+// import * as generateCommand from './commands/generate'
+// import * as infoCommand from './commands/info'
+// import * as lintCommand from './commands/lint'
+// import * as prerenderCommand from './commands/prerender'
+// import * as prismaCommand from './commands/prisma'
+// import * as recordCommand from './commands/record'
+// import * as serveCommand from './commands/serve'
+// import * as setupCommand from './commands/setup'
+// import * as storybookCommand from './commands/storybook'
+// import * as testCommand from './commands/test'
+// import * as tstojsCommand from './commands/ts-to-js'
+// import * as typeCheckCommand from './commands/type-check'
+// import * as upgradeCommand from './commands/upgrade'
 
 /**
  * The current working directory can be set via:
@@ -64,14 +86,34 @@ const loadDotEnvDefaultsMiddleware = () => {
 yargs
   .scriptName('rw')
   .middleware([
-    getCwdMiddleware,
-    loadDotEnvDefaultsMiddleware,
-    telemetryMiddleware,
+    // getCwdMiddleware,
+    // loadDotEnvDefaultsMiddleware,
+    // telemetryMiddleware,
   ])
   .option('cwd', {
     describe: 'Working directory to use (where `redwood.toml` is located.)',
   })
-  .commandDir('./commands')
+  .command(buildCommand)
+  // .command(checkCommand)
+  .command(consoleCommand)
+  // .command(dataMigrateCommand)
+  // .command(deployCommand)
+  // .command(destroyCommand)
+  .command(devCommand)
+  // .command(execCommand)
+  // .command(generateCommand)
+  // .command(infoCommand)
+  // .command(lintCommand)
+  // .command(prerenderCommand)
+  // .command(prismaCommand)
+  // .command(recordCommand)
+  // .command(serveCommand)
+  // .command(setupCommand)
+  // .command(storybookCommand)
+  // .command(testCommand)
+  // .command(tstojsCommand)
+  // .command(typeCheckCommand)
+  // .command(upgradeCommand)
   .example(
     'yarn rw g page home /',
     "\"Create a page component named 'Home' at path '/'\""
